+++ upsetplot/tests/test_upsetplot.py
@@ -2,7 +2,9 @@ import io
 import itertools
 
 import pytest
-from pandas.util.testing import assert_series_equal
+from pandas.util.testing import (
+    assert_series_equal, assert_frame_equal, assert_index_equal)
+from numpy.testing import assert_array_equal
 import pandas as pd
 import numpy as np
 import matplotlib.figure
@@ -25,10 +27,16 @@ def is_ascending(seq):
 ])
 @pytest.mark.parametrize('sort_by', ['cardinality', 'degree'])
 @pytest.mark.parametrize('sort_sets_by', [None, 'cardinality'])
-def test_process_data(X, sort_by, sort_sets_by):
-    intersections, totals = _process_data(X,
-                                          sort_by=sort_by,
-                                          sort_sets_by=sort_sets_by)
+def test_process_data_series(X, sort_by, sort_sets_by):
+    with pytest.raises(ValueError, match='sum_over is not applicable'):
+        _process_data(X, sort_by=sort_by, sort_sets_by=sort_sets_by,
+                      sum_over=False)
+
+    df, intersections, totals = _process_data(X,
+                                              sort_by=sort_by,
+                                              sort_sets_by=sort_sets_by,
+                                              sum_over=None)
+    assert intersections.name == 'value'
     X_reordered = (X
                    .reorder_levels(intersections.index.names)
                    .reindex(index=intersections.index))
@@ -49,20 +57,103 @@ def test_process_data(X, sort_by, sort_sets_by):
 
     assert np.all(totals.index.values == intersections.index.names)
 
+    assert np.all(df.index.names == intersections.index.names)
+    assert set(df.columns) == {'_value', '_bin'}
+    assert_index_equal(df['_value'].reorder_levels(X.index.names).index,
+                       X.index)
+    assert_array_equal(df['_value'], X)
+    assert_index_equal(intersections.iloc[df['_bin']].index,
+                       df.index)
+    assert len(df) == len(X)
+
+
+@pytest.mark.parametrize('x', [
+    generate_data(aggregated=False),
+])
+@pytest.mark.parametrize('sort_by', ['cardinality', 'degree'])
+@pytest.mark.parametrize('sort_sets_by', [None, 'cardinality'])
+def test_process_data_frame(x, sort_by, sort_sets_by):
+    X = pd.DataFrame({'a': x})
+
+    with pytest.raises(ValueError, match='sum_over must be False or '):
+        _process_data(X, sort_by=sort_by, sort_sets_by=sort_sets_by,
+                      sum_over=None)
+
+    df, intersections, totals = _process_data(X,
+                                              sort_by=sort_by,
+                                              sort_sets_by=sort_sets_by,
+                                              sum_over='a')
+    assert df is not X
+
+    # check equivalence to Series
+    df1, intersections1, totals1 = _process_data(x,
+                                                 sort_by=sort_by,
+                                                 sort_sets_by=sort_sets_by,
+                                                 sum_over=None)
+
+    assert intersections.name == 'a'
+    assert_frame_equal(df, df1.rename(columns={'_value': 'a'}))
+    assert_series_equal(intersections, intersections1, check_names=False)
+    assert_series_equal(totals, totals1)
+
+    # check effect of extra column
+    X = pd.DataFrame({'a': x, 'b': np.arange(len(x))})
+    df2, intersections2, totals2 = _process_data(X,
+                                                 sort_by=sort_by,
+                                                 sort_sets_by=sort_sets_by,
+                                                 sum_over='a')
+    assert_series_equal(intersections, intersections2)
+    assert_series_equal(totals, totals2)
+    assert_frame_equal(df, df2.drop('b', axis=1))
+    assert_array_equal(df2['b'], X['b'])  # disregard levels, tested above
+
+    # check effect not dependent on order/name
+    X = pd.DataFrame({'b': np.arange(len(x)), 'c': x})
+    df3, intersections3, totals3 = _process_data(X,
+                                                 sort_by=sort_by,
+                                                 sort_sets_by=sort_sets_by,
+                                                 sum_over='c')
+    assert_series_equal(intersections, intersections3, check_names=False)
+    assert intersections.name == 'a'
+    assert intersections3.name == 'c'
+    assert_series_equal(totals, totals3)
+    assert_frame_equal(df.rename(columns={'a': 'c'}), df3.drop('b', axis=1))
+    assert_array_equal(df3['b'], X['b'])
+
+    # check sum_over=False
+    X = pd.DataFrame({'b': np.ones(len(x), dtype=int), 'c': x})
+    df4, intersections4, totals4 = _process_data(X,
+                                                 sort_by=sort_by,
+                                                 sort_sets_by=sort_sets_by,
+                                                 sum_over='b')
+    df5, intersections5, totals5 = _process_data(X,
+                                                 sort_by=sort_by,
+                                                 sort_sets_by=sort_sets_by,
+                                                 sum_over=False)
+    assert_series_equal(intersections4, intersections5, check_names=False)
+    assert intersections4.name == 'b'
+    assert intersections5.name == 'size'
+    assert_series_equal(totals4, totals5)
+    assert_frame_equal(df4, df5)
+
 
 @pytest.mark.parametrize('sort_by', ['cardinality', 'degree'])
 @pytest.mark.parametrize('sort_sets_by', [None, 'cardinality'])
 def test_not_aggregated(sort_by, sort_sets_by):
     # FIXME: this is not testing if aggregation used is count or sum
-    kw = {'sort_by': sort_by, 'sort_sets_by': sort_sets_by}
+    kw = {'sort_by': sort_by, 'sort_sets_by': sort_sets_by, 'sum_over': None}
     Xagg = generate_data(aggregated=True)
-    intersections1, totals1 = _process_data(Xagg, **kw)
+    df1, intersections1, totals1 = _process_data(Xagg, **kw)
     Xunagg = generate_data()
     Xunagg.loc[:] = 1
-    intersections2, totals2 = _process_data(Xunagg, **kw)
+    df2, intersections2, totals2 = _process_data(Xunagg, **kw)
     assert_series_equal(intersections1, intersections2,
                         check_dtype=False)
     assert_series_equal(totals1, totals2, check_dtype=False)
+    assert set(df1.columns) == {'_value', '_bin'}
+    assert set(df2.columns) == {'_value', '_bin'}
+    assert len(df2) == len(Xunagg)
+    assert df2['_bin'].nunique() == len(intersections2)
 
 
 @pytest.mark.parametrize('kw', [{'sort_by': 'blah'},
@@ -110,7 +201,7 @@ def test_dataframe_raises():
     df = pd.DataFrame({'val': [5, 7],
                        'set1': [False, True],
                        'set2': [True, True]}).set_index(['set1', 'set2'])
-    with pytest.raises(ValueError, match='must be a pandas.Series'):
+    with pytest.raises(ValueError, match='sum_over must be'):
         plot(df, fig)
 
 
@@ -190,3 +281,38 @@ def test_show_counts(orientation):
     with pytest.raises(ValueError):
         fig = matplotlib.figure.Figure()
         plot(X, fig, show_counts='%0.2h')
+
+
+def test_add_catplot():
+    pytest.importorskip('seaborn')
+    X = generate_data(n_samples=100)
+    upset = UpSet(X)
+    # smoke test
+    upset.add_catplot('violin')
+    fig = matplotlib.figure.Figure()
+    upset.plot(fig)
+
+    # can't provide value with Series
+    with pytest.raises(ValueError):
+        upset.add_catplot('violin', value='foo')
+
+    # check the above add_catplot did not break the state
+    upset.plot(fig)
+
+    X = generate_data(n_samples=100)
+    X.name = 'foo'
+    X = X.to_frame()
+    upset = UpSet(X, sum_over=False)
+    # must provide value with DataFrame
+    with pytest.raises(ValueError):
+        upset.add_catplot('violin')
+    upset.add_catplot('violin', value='foo')
+    with pytest.raises(ValueError):
+        # not a known column
+        upset.add_catplot('violin', value='bar')
+    upset.plot(fig)
+
+    # invalid plot kind raises error when plotting
+    upset.add_catplot('foobar', value='foo')
+    with pytest.raises(AttributeError):
+        upset.plot(fig)